// src/modules/booking/booking.enum.ts
export enum SORT_VALUE {
    TODAY = 'today',
    THIS_WEEK = 'this_week',
    THIS_MONTH = 'this_month',
    LAST_MONTH = 'this_month',
    CONFIRMED = 'confirmed',
    EXPIRED = 'expired',
    CANCELED = 'canceled',
    SCHEDULED = 'scheduled',
    DELETED = 'deleted'
  }
  
  export enum BOOKING_STATUS {
    CONFIRMED = 'confirmed',
    EXPIRED = 'expired',
    CANCELED = 'canceled',
    SCHEDULED = 'scheduled',
  }